requirements.txt
ase==3.25.0
contourpy==1.3.2
cycler==0.12.1
dscribe==2.1.1
fonttools==4.59.0
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
numba==0.61.2
numpy==2.2.6
ovito==3.13.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
shiboken6==6.9.1
six==1.17.0
sparse==0.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
traits==7.0.2
tzdata==2025.2
